requirements.txt
pytest==7.4.4
google-cloud-bigquery[pandas]==2.34.3
Flask==2.2.2
Werkzeug==2.2.2
flask-restx==1.1.0